pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain-community>=0.3.19",
    "langchain>=0.3.20",
    "numpy>=2.2.3",
    "openai>=1.65.3",
    "pypdf2>=3.0.1",
    "streamlit>=1.43.0",
    "tiktoken>=0.9.0",
]
